modify dist_utils
