+++ easycv/models/detection/utils/__init__.py
@@ -3,9 +3,5 @@
 
 from .boxes import (bbox2result, bboxes_iou, box_cxcywh_to_xyxy,
                     box_xyxy_to_cxcywh, generalized_box_iou, postprocess, bbox_overlaps, distance2bbox)
-from .dist_utils import reduce_mean
-from .instances import Boxes, Instances
-from .misc import (accuracy, get_world_size, interpolate,
-                   is_dist_avail_and_initialized, multi_apply,
-                   nested_tensor_from_tensor_list, select_single_mlvl, filter_scores_and_topk)
+from .misc import (accuracy, interpolate, multi_apply, select_single_mlvl, filter_scores_and_topk)
 from .utils import output_postprocess
